Added floating point number support

+++ main.py
@@ -1,5 +1,5 @@
-num1 = int(input("Enter a number: "))
-num2 = int(input("Enter another number: "))
+num1 = float(input("Enter a number: "))
+num2 = float(input("Enter another number: "))
 
 result = num1 + num2
